RSA tuning in progress.

+++ compile.sh
@@ -5,8 +5,14 @@ set -e
 if [ ! -d "build" ]; then
     mkdir build
 fi
-rm -rd dbg
+if [ -d "dbg" ]; then
+    rm -rd dbg
+fi
+if [ -d "tmp" ]; then
+    rm -rd tmp
+fi
 mkdir dbg
+mkdir tmp
 cd build
 cmake .. $@
 make $MAKEOPTS
